+++ CHANGELOG.md
@@ -1,3 +1,2 @@
 ## Bugfixes
-- Fixed modded barriers only being able to be seen by operators/players with permission level 2.
-- Fixed Mob related predicate barriers being able to be bypassed with vehicles (such as boats and minecarts).
+- Advanced Barriers used a double prefix in the icon path.